refactor(utils): card에서 사용되는 함수 유틸로 분리
- card컴포넌트가 다른곳에서도 사용되므로, utils에 함수들을 분리시킵니다.

#50
